added my scene with game-view
added two new scenes. One with my name, and a test scene for NPC movement.

In my scene, I've altered the game view to allow for a more interesting, "flat" camera angle that focuses on the action in the Reaper's immediate surroundings.

Rahul will create a short script to follow the player