Creation of Canvas Menus prefabs and implementation in Level 01 testing scene
Canvas: In-Game, Pause-Game and End-Game created and set up with the End-Game Panel and Number of NPCs killed on the Game Manager.
There is a message that can be set up each time that a NPC is dead and points at the bottom of the screen to give feedback for the player.
Synced Quit event for Exit buttons to the Quit game fuctions on Game Manager script too.
And finally, creation of a Main Menu.